refactor : 封装页面刷行到 hook
